% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{carry_price_forward}
\alias{carry_price_forward}
\title{Carry prices forward}
\usage{
carry_price_forward(x)
}
\arguments{
\item{x}{an xts object that may have NA's}
}
\value{
an xts object where NA's are replace by the last valid value. Note
        that if NA's appear in the beginning of the series they will appear
        in the result.
}
\description{
Carry prices forward
}

