% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cointegration.r
\name{cointegration_info}
\alias{cointegration_info}
\title{Get cointegration information for time series data}
\usage{
cointegration_info(x)
}
\arguments{
\item{x}{a data.frame, matrix, or xts object}
}
\value{
a list containting the pairwise cointegration p-values, the
        p-value of the Kolmogorov-Smirnov test for uniformity in the p-values
        (null is uniformity), and a measure of the "cointegratedness" of
        the time series vectors (100 indicates all are cointegrated zero
        indicates they are all independent).
}
\description{
Get cointegration information for time series data
}

