% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cointegration.r
\name{cointegration_p_matrix}
\alias{cointegration_p_matrix}
\title{Get the pairwise matrix of p-values testing for cointegration}
\usage{
cointegration_p_matrix(x)
}
\arguments{
\item{x}{a data.frame, matrix, or xts object}
}
\value{
An upper triangular matrix where each entry corresponds to the
p-value from McKinnons's test assuming no time trend. Note that if one
stock in a pair-wise cointegration is constant the return value will be 1.
}
\description{
Get the pairwise matrix of p-values testing for cointegration
}
\examples{
data(aapl_fix)
data(a_fix)
aapl_cons = consolidate_prices(aapl_fix$sys_date, aapl_fix$sys_time,
 aapl_fix$td_price, aapl_fix$td_size)
a_cons = consolidate_prices(a_fix$sys_date, a_fix$sys_time,
 a_fix$td_price, a_fix$td_size)
trades = merge(aapl_cons$price, a_cons$price)
trades = carry_prices_forward(trades)
names(trades) = c("aapl", "a")
cointegration_p_matrix(trades)
}

