% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process.R
\name{consolidate_taq_trades}
\alias{consolidate_taq_trades}
\title{Consolidate trades at a single second resolution}
\usage{
consolidate_taq_trades(x)
}
\arguments{
\item{x}{a data frame with columns time price and size}
}
\value{
an xts matrix with price, max_price, min_price, and vwap prices
}
\description{
Consolidate trades at a single second resolution
}

