% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nxcore.R
\docType{data}
\name{exg_lookup}
\alias{exg_lookup}
\title{Exchange Lookup}
\source{
\url{http://nxcoreapi.com/doc/table_NxST_EXCHANGE.html}
}
\usage{
exg_lookup
}
\description{
Exchange lookup
}
\keyword{data}

