% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitops.R
\name{get_bit_matrix}
\alias{get_bit_matrix}
\title{Convert decimal to matrix of bits}
\usage{
get_bit_matrix(x, n = 32)
}
\arguments{
\item{x}{decimal or vector of decimal integers}

\item{n}{number of bits}
}
\description{
Convert decimal to matrix of bits
}
\examples{
get_bit_matrix(1:10)
get_bit_matrix(1:10, n = 4)
}

