% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bitops.R
\name{get_td_cond_flags}
\alias{get_td_cond_flags}
\title{Convert decimal NxCore trade condition flags to bit representation}
\usage{
get_td_cond_flags(x)
}
\arguments{
\item{x}{vector of decimal trade condition flag values}
}
\description{
Convert decimal NxCore trade condition flags to bit representation
}
\examples{
get_td_cond_flags(1)
}

