% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bitops.R
\name{get_td_price_flags}
\alias{get_td_price_flags}
\title{Convert decimal NxCore price flags to bit representation}
\usage{
get_td_price_flags(x)
}
\arguments{
\item{x}{vector of decimal price flag values}
}
\description{
Convert decimal NxCore price flags to bit representation
}
\examples{
get_td_price_flags(1)
}

