% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{log_rate_return}
\alias{log_rate_return}
\title{Get the log-rate-returns of a time-series}
\usage{
log_rate_return(prices)
}
\arguments{
\item{prices}{the values to calculated the log-rate returns from.}
}
\description{
Get the log-rate-returns of a time-series
}

