% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nxcore.R
\docType{data}
\name{nxcore_dict}
\alias{nxcore_dict}
\title{NxCore Data Dictionary}
\usage{
nxcore_dict
}
\description{
List of variable names and labels for each of the four types of data (trade, etc.)
}
\keyword{data}

