% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process.R
\name{process_insertions}
\alias{process_insertions}
\title{Take care of insertions in NxCore trade data}
\usage{
process_insertions(x)
}
\arguments{
\item{x}{a data frame for a single day and a single symbol of NxCore trade data, as read by \code{\link{read_nx_trade}}}
}
\description{
Take care of insertions in NxCore trade data
}
\note{
The approach is ...
}

