% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read_nx_trade}
\alias{read_nx_trade}
\title{Turn XDATA-processed NxCore "trade" text file into R data frame}
\usage{
read_nx_trade(file)
}
\description{
Turn XDATA-processed NxCore "trade" text file into R data frame
}
\note{
This presumes the input is a single file
}

