% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/windowed_metrics.r
\name{realized_volatility}
\alias{realized_volatility}
\title{Calculate the realized volatility for a transction price time-series}
\usage{
realized_volatility(price, interval = minutes(5), interval_const = 1)
}
\arguments{
\item{price}{and xts object containting transaction prices}

\item{interval}{the window size to calculate realized volatility over. The
default is 5 mintes.}

\item{interval}{const the number of trading periods specified by the interval
is the number of trading days in a year. It should be noted that this is
only needed for calculations like realized volatility. For comparison
of two stocks over the same intervals it is fine to leave this as 1.}
}
\description{
Calculate the realized volatility for a transction price time-series
}
\examples{
data(aapl_fix)
ctp = consolidate_prices(aapl_fix$sys_date, aapl_fix$sys_time,
 aapl_fix$td_price, aapl_fix$td_size, date_format="\%Y-\%m-\%d",
 time_format="\%H:\%M:\%S")
rv = realized_volatility(ctp$price)
plot(rv)
}

