% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process.R
\name{register_prices}
\alias{register_prices}
\title{Register price information across a set of stocks.}
\usage{
register_prices(x, on = "seconds", k = 1)
}
\arguments{
\item{x}{the list of stocks with a column "price"}

\item{on}{the resolution to consolidate trades. The default is "seconds"
and any resolution supported by the xts::endpoints function is valid.}

\item{k}{along every k-th elements. For example if "seconds" is specified
as the "on" parameter and k=2 then prices are consolidated for every 2
seconds.}
}
\description{
Takes a list of stocks from the output of consolidate prices,
registers them in time, and display them at the specified resolution.
}

