% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nxcore.R
\docType{data}
\name{russell_or_sp}
\alias{russell_or_sp}
\title{Russell 1000 and SP 500 Stocks}
\usage{
russell_or_sp
}
\description{
Ticker symbols for stocks in the Russell 1000 and SP 500
}
\keyword{data}

