% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nxcore.R
\docType{data}
\name{symb_b}
\alias{symb_b}
\alias{symb_e}
\alias{symb_f}
\alias{symb_i}
\alias{symb_m}
\alias{symb_p}
\alias{symb_z}
\title{Symbol name lookup tables}
\usage{
symb_b, symb_e, symb_f, symb_i, symb_m, symb_p, symb_z
}
\description{
Tables used to look up symbol names for different security types
}
\keyword{data}

