% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nxcore.R
\docType{data}
\name{symbol_prefix_lookup}
\alias{symbol_prefix_lookup}
\title{Symbol Prefix Lookup}
\source{
\url{http://nxcoreapi.com/doc/info_SymbolPrefixes.html}
}
\usage{
symbol_prefix_lookup
}
\description{
Symbol prefix lookup
}
\keyword{data}

