% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nxcore.R
\docType{data}
\name{td_cond_flag_lookup}
\alias{td_cond_flag_lookup}
\title{Trade Condition Flag Lookup}
\source{
\url{http://nxcoreapi.com/doc/struct_NxCoreTrade.html#conditionflags}
}
\usage{
td_cond_flag_lookup
}
\description{
Trade condition flag lookup
}
\keyword{data}

