% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nxcore.R
\docType{data}
\name{td_cond_idx_lookup}
\alias{td_cond_idx_lookup}
\title{Trade Condition Index Lookup}
\source{
\url{http://nxcoreapi.com/doc/info_TradeConditions.html}
}
\usage{
td_cond_idx_lookup
}
\description{
Trade condition index lookup
}
\keyword{data}

