% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nxcore.R
\docType{data}
\name{td_price_flag_lookup}
\alias{td_price_flag_lookup}
\title{Trade Price Flag Lookup}
\source{
\url{http://nxcoreapi.com/doc/struct_NxCoreTrade.html#PriceFlags}
}
\usage{
td_price_flag_lookup
}
\description{
Trade price flag lookup
}
\keyword{data}

