% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nxcore.R
\docType{data}
\name{td_vol_type_lookup}
\alias{td_vol_type_lookup}
\title{Trade Volume Type Lookup}
\source{
\url{http://nxcoreapi.com/doc/struct_NxCoreTrade.html#volumetype}
}
\usage{
td_vol_type_lookup
}
\description{
Trade volume type lookup
}
\keyword{data}

