% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cointegration.r
\name{cointegration_p_matrix}
\alias{cointegration_p_matrix}
\title{Get the pairwise matrix of p-values testing for cointegration}
\usage{
cointegration_p_matrix(x)
}
\arguments{
\item{x}{a data.frame, matrix, or xts object}
}
\value{
An upper triangular matrix where each entry corresponds to the result
        of McKinnons's test assuming no intercept or time trend.
}
\description{
Get the pairwise matrix of p-values testing for cointegration
}

