% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/process.R
\name{consolidate_prices}
\alias{consolidate_prices}
\title{Consolidate prices at a single second resolution}
\usage{
consolidate_prices(x)
}
\arguments{
\item{x}{a data frame with columns named including sys_date, sys_time,
td_price, td_size.}
}
\value{
an xts matrix with (vwapped) price, max price, min price, and volume.
}
\description{
Consolidate prices at a single second resolution
}
\examples{
# Consolidate TAQ trades...
data(aapl_taq)
ctp = consolidate_prices(aapl_taq)
# or FIX trades...
data(aapl_fix)
cfp = consolidate_prices(aapl_fix)
}

