% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.r
\name{vwap}
\alias{vwap}
\title{Calculated the VWAP for a specified interval}
\usage{
vwap(price, size, date_time, on = "minutes", k = 1)
}
\arguments{
\item{price}{the product's tranaction price.}

\item{size}{the number of products transacted.}

\item{on}{the interval to find the average over.}

\item{k}{along every k elements.}
}
\value{
an xts object with the volume-weighted average price for the
        specified interval of time.
}
\description{
Calculated the VWAP for a specified interval
}

